Fix FPS limit setting type conversion

+++ D3D11Engine/GothicAPI.cpp
@@ -5498,8 +5498,8 @@ XRESULT GothicAPI::LoadMenuSettings( const std::string& file ) {
         s.EnableOcclusionCulling = GetPrivateProfileBoolA( "General", "EnableOcclusionCulling", ds.EnableOcclusionCulling, ini );
         s.FpsLimit = GetPrivateProfileIntA( "General", "FpsLimit", 0, ini.c_str() );
         s.FpsLimitLastEnabled = std::clamp(
-            GetPrivateProfileIntA( "General", "FpsLimitLastEnabled",
-                s.FpsLimit > 0 ? s.FpsLimit : ds.FpsLimitLastEnabled, ini.c_str() ),
+            static_cast<int>(GetPrivateProfileIntA( "General", "FpsLimitLastEnabled",
+                s.FpsLimit > 0 ? s.FpsLimit : ds.FpsLimitLastEnabled, ini.c_str() )),
             10, 300 );
         if ( s.FpsLimit > 0 ) {
             s.FpsLimitLastEnabled = std::clamp( s.FpsLimit, 10, 300 );
